+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Unreleased
+# 3.16.0
 * Add `LiabilityShift` class and `getLiabilityShift` to RiskData
 * Fixed a bug where `decision reasons` was returning an un-parsed string instead of array